requirements.txt
pytest>=7.1.2
selene>=2.0.0b8
allure-pytest==2.9.45
python-dotenv==0.20.0
selenium==4.3.0